+++ src/main/java/com/vlkan/rfos/RotatingFileOutputStream.java
@@ -16,13 +16,19 @@
 
 package com.vlkan.rfos;
 
+import com.vlkan.rfos.policy.RotationPolicy;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
+import java.nio.file.Files;
 import java.nio.file.Paths;
+import java.nio.file.StandardCopyOption;
 import java.time.Instant;
 import java.util.ArrayList;
 import java.util.List;
@@ -30,11 +36,6 @@ import java.util.Objects;
 import java.util.Set;
 import java.util.function.Consumer;
 import java.util.zip.GZIPOutputStream;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import com.vlkan.rfos.policy.RotationPolicy;
 
 public class RotatingFileOutputStream extends OutputStream implements Rotatable {
 
@@ -116,27 +117,18 @@ public class RotatingFileOutputStream extends OutputStream implements Rotatable 
         invokeCallbacks(callback -> callback.onClose(policy, instant, stream));
         stream.close();
 
-        // Rename backups, if enabled.
+        // Backup file, if enabled.
         File rotatedFile;
-        boolean renamed;
         if (config.getMaxBackupCount() > 0) {
-            File[] rotatedFileRef = {null};
-            renamed = renameBackups() && backupFile(rotatedFileRef);
-            rotatedFile = rotatedFileRef[0];
+            renameBackups();
+            rotatedFile = backupFile();
         }
 
         // Otherwise, rename using the provided file pattern.
         else {
         	rotatedFile = config.getFilePattern().create(instant).getAbsoluteFile();
             LOGGER.debug("renaming {file={}, rotatedFile={}}", config.getFile(), rotatedFile);
-            renamed = config.getFile().renameTo(rotatedFile);
-        }
-
-        if (!renamed) {
-            String message = String.format("rename failure {file=%s, rotatedFile=%s}", config.getFile(), rotatedFile);
-            IOException error = new IOException(message);
-            invokeCallbacks(callback -> callback.onFailure(policy, instant, rotatedFile, error));
-            return;
+            renameFile(config.getFile(), rotatedFile);
         }
 
         // Re-open the file.
@@ -154,7 +146,7 @@ public class RotatingFileOutputStream extends OutputStream implements Rotatable 
 
     }
 
-    private boolean renameBackups() {
+    private void renameBackups() throws IOException {
         File dstFile = getBackupFile(config.getMaxBackupCount() - 1);
         for (int backupIndex = config.getMaxBackupCount() - 2; backupIndex >= 0; backupIndex--) {
             File srcFile = getBackupFile(backupIndex);
@@ -162,21 +154,26 @@ public class RotatingFileOutputStream extends OutputStream implements Rotatable 
                 continue;
             }
             LOGGER.debug("renaming backup {srcFile={}, dstFile={}}", srcFile, dstFile);
-            boolean renamed = srcFile.renameTo(dstFile);
-            if (!renamed) {
-                LOGGER.error("failed renaming backup {srcFile={}, dstFile={}}", srcFile, dstFile);
-                return false;
-            }
+            renameFile(srcFile, dstFile);
             dstFile = srcFile;
         }
-        return true;
     }
 
-    private boolean backupFile(File[] backupFileRef) {
-        File dstFile = backupFileRef[0] = getBackupFile(0);
+    private File backupFile() throws IOException {
+        File dstFile = getBackupFile(0);
         File srcFile = config.getFile();
         LOGGER.debug("renaming for backup {srcFile={}, dstFile={}}", srcFile, dstFile);
-        return srcFile.renameTo(dstFile);
+        renameFile(srcFile, dstFile);
+        return dstFile;
+    }
+
+    private static void renameFile(File srcFile, File dstFile) throws IOException {
+        Files.move(
+                srcFile.toPath(),
+                dstFile.toPath(),
+                StandardCopyOption.REPLACE_EXISTING/*,      // The rest of the arguments (atomic & copy-attr) are pretty
+                StandardCopyOption.ATOMIC_MOVE,             // much platform-dependent and JVM throws an "unsupported
+                StandardCopyOption.COPY_ATTRIBUTES*/);      // option" exception at runtime.
     }
 
     private File getBackupFile(int backupIndex) {